Moved Related-Posts to Components
